Refactor bloodlines actions and reducers to a common util files
