+++ src/reducers/configureReducers.js
@@ -1,10 +1,11 @@
-import { combineReducers } from 'redux';
+import {combineReducers} from 'redux';
 
-import { warehouseReducer } from './warehouseReducer';
-import { roasterReducer } from './roasterReducer';
-import { userReducer } from './userReducer';
-import { customerPaymentInfoReducer } from './coinageReducer';
-import { triggers, contents, receipts, modify } from './bloodlinesReducer';
+import {warehouseReducer} from './warehouseReducer';
+import {roasterReducer} from './roasterReducer';
+import {userReducer} from './userReducer';
+import {customerPaymentInfoReducer} from './coinageReducer';
+import {triggers, contents, receipts} from './bloodlinesReducer';
+import ReducerUtil from './reducerUtil';
 
 const rootReducer = combineReducers({
     roasterReducer,
@@ -14,7 +15,7 @@ const rootReducer = combineReducers({
     triggers,
     contents,
     receipts,
-    modify
+    modify: ReducerUtil.modify
 });
 
 export default rootReducer;
